Move `json_modules` out of `api`
