+++ src/json_models/chat_completion.rs
